+++ srcCpp03/perftest_cpp.h
@@ -73,7 +73,7 @@ class perftest_cpp
     }
 
   private:
-    int  _DataLen;
+    unsigned long _DataLen;
     int  _BatchSize;
     int  _SamplesPerBatch;
     unsigned long long _NumIter;
@@ -94,7 +94,7 @@ class perftest_cpp
     int _pubRate;
     bool _pubRateMethodSpin;
     bool _isKeyed;
-    int _useUnbounded;
+    unsigned long _useUnbounded;
     unsigned int _executionTime;
     bool _displayWriterStats;
